# clinical/clinical_service/api/swagger.yaml
swagger: '2.0'
info:
  title: Clinical query demo - SQL wrapper
  version: "1"
basePath: /clinical/v1
consumes:
  - application/json
produces:
  - application/json

paths:
  /individuals/{individual_id}:
    get:
      operationId: clinical_service.api.operations.get_one_individual
      summary: Get specific individual
      parameters:
        - $ref: '#/parameters/individual_id'
      responses:
        "200":
          description: Return individual
          schema:
            $ref: '#/definitions/Individual'
        "401":
          description: Unauthenticated
          schema:
            $ref: "#/definitions/Error"
        "403":
          description: Forbidden
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: Internal error
          schema:
            $ref: "#/definitions/Error"


  /individuals:
    get:
      operationId: clinical_service.api.operations.get_individuals
      summary: Get individuals data from DB
      parameters:
        - in: query
          name: whereConditions
          type: array
          required: false
          collectionFormat: csv
          items:
            type: string
      responses:
        "200":
          description: Return individuals
          schema:
            type: array
            example: []
            items:
              $ref: '#/definitions/Individual'
        "401":
          description: Unauthenticated
          schema:
            $ref: "#/definitions/Error"
        "403":
          description: Forbidden
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: Internal error
          schema:
            $ref: "#/definitions/Error"


parameters:
  individual_id:
    name: individual_id
    description: Individual ID
    in: path
    type: string
    x-example: P0001
    required: true


definitions:
  Individual:
    type: object
    required:
      - status
    properties:
      id:
        type: string
        description: Participant ID
        example: "P0001"
        readOnly: true
        minLength: 1
        maxLength: 100
      status:
        type: string
        description: description of individual
        example: "Healthy"
        minLength: 1
        maxLength: 100

  Error:
    type: object
    required:
      - message
      - code
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string